backend/requirements.txt
amqp==5.3.1
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.5
daphne==4.2.1
Django==5.2.12
django-auditlog==3.4.1
django-celery-beat==2.9.0
django_celery_results==2.6.0
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-timezone-field==7.2.1
dnspython==2.8.0
email-validator==2.3.0
google-api-core==2.30.0
google-api-python-client==2.192.0
googleapis-common-protos==1.73.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
kombu==5.6.2
msgpack==1.1.2
oauthlib==3.3.1
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.5
proto-plus==1.27.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
py-ubjson==0.16.1
redis==7.3.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
zope.interface==8.2
